requirements.txt
pandas>=2.2
scikit-learn>=1.2
lightgbm>=4.0
statsmodels>=0.14
entsoe-py>=0.7
python-dateutil
joblib
matplotlib
shap
pyarrow>=10.0
fastparquet>=0.8
requests>=2.31
python-dotenv>=1.0
